Separate token files
