+++ packages/__cxstarter__/src/styleguide/index.ts
@@ -1,2 +1,2 @@
 export { __cxstarter__StyleGuidePage } from './Page';
-export { __cxstarter__StyleGuideTemplate } from './StyleGuideTemplate';
+// export { __cxstarter__StyleGuideTemplate } from './StyleGuideTemplate';
